+++ src/AcuteBenchmark.jl
@@ -1,5 +1,6 @@
 module AcuteBenchmark
 
+include("utils.jl")
 include("benchmarks.jl")
 # include("io.jl")
 include("plots.jl")
